Many various additions and improvements
Optimized footprint calculation
Added ReplanParams and PlannerStats classes
Added ways to call replan
Added main function and command-line interface
Made functions in base planner class polymorphic
Added test script to run through various SBPL examples
Generated 10cm PR2 motion primitives
Moved old SVN history into new Git repo
Changed all SBPL_PRINTFs to normal printfs in main

+++ src/discrete_space_information/robarm/environment_robarm.cpp
@@ -127,7 +127,7 @@ EnvROBARMHashEntry_t* EnvironmentROBARM::GetHashEntry(short unsigned int* coord,
 	if ((int)EnvROBARM.Coord2StateIDHashTable[binid].size() > 500)
 	{
 		SBPL_PRINTF("WARNING: Hash table has a bin %d (coord0=%d) of size %d\n", 
-			binid, coord[0], EnvROBARM.Coord2StateIDHashTable[binid].size());
+			binid, coord[0], (int)EnvROBARM.Coord2StateIDHashTable[binid].size());
 		
 		PrintHashTableHist();		
 	}
